backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
httpx==0.24.1
sqlalchemy==2.0.20
pymysql==1.1.0
cryptography==41.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
beautifulsoup4==4.12.2
cachetools==5.3.1
firebase-admin==6.2.0
aiohttp==3.8.5
